# docs/swagger.yaml
...
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  pager.PageModel:
    properties:
      count:
        type: integer
      list: {}
    type: object
...
  schema.GetRankPersonalPageResp:
    properties:
      answer_id:
        description: answer id
        type: string
      content:
        description: content
        type: string
      created_at:
        description: create time
        type: integer
      object_id:
        description: object id
        type: string
      object_type:
        description: object type
        enum:
        - question
        - answer
        - tag
        - comment
        type: string
      question_id:
        description: question id
        type: string
      rank_type:
        description: rank type
        type: string
      reputation:
        description: reputation
        type: integer
      title:
        description: title
        type: string
      url_title:
        description: url title
        type: string
    type: object
...
  schema.GetSiteLegalInfoResp:
    properties:
      privacy_policy_original_text:
        type: string
      privacy_policy_parsed_text:
        type: string
      terms_of_service_original_text:
        type: string
      terms_of_service_parsed_text:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /answer/api/v1/language/config:
    get:
      description: get language config mapping
      parameters:
      - description: Accept-Language
        in: header
        name: Accept-Language
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      summary: get language config mapping
      tags:
      - Lang
...
  /answer/api/v1/permission:
    get:
      description: check user permission
      parameters:
      - description: access-token
        in: header
        name: Authorization
        required: true
        type: string
      - description: permission key
        enum:
        - question.add
        - question.edit
        - question.edit_without_review
        - question.delete
        - question.close
        - question.reopen
        - question.vote_up
        - question.vote_down
        - question.pin
        - question.unpin
        - question.hide
        - question.show
        - answer.add
        - answer.edit
        - answer.edit_without_review
        - answer.delete
        - answer.accept
        - answer.vote_up
        - answer.vote_down
        - answer.invite_someone_to_answer
        - comment.add
        - comment.edit
        - comment.delete
        - comment.vote_up
        - comment.vote_down
        - report.add
        - tag.add
        - tag.edit
        - tag.edit_slug_name
        - tag.edit_without_review
        - tag.delete
        - tag.synonym
        - link.url_limit
        - vote.detail
        - answer.audit
        - question.audit
        - tag.audit
        - tag.use_reserved_tag
        in: query
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  additionalProperties:
                    type: boolean
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: check user permission
      tags:
      - Permission
...
  /answer/api/v1/personal/rank/page:
    get:
      description: user personal rank list
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: username
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/pager.PageModel'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/schema.GetRankPersonalPageResp'
                        type: array
                    type: object
              type: object
      summary: user personal rank list
      tags:
      - Rank
...
  /answer/api/v1/siteinfo/legal:
    get:
      description: get site legal info
      parameters:
      - description: legal information type
        enum:
        - tos
        - privacy
        in: query
        name: info_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetSiteLegalInfoResp'
              type: object
      summary: get site legal info
      tags:
      - site
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"